updated data for test_tabulate_path - there was a bug because it was no external terms in conv II
