+++ tests/test_path.py
@@ -1,6 +1,7 @@
 import os
 import pickle
 import numpy as np
+import pytest
 from pytest import approx
 
 import wannierberri as wberri
@@ -144,3 +145,22 @@ def test_tabulate_path(system_Haldane_PythTB, check_run):
         fatfactor=20,
         cut_k=True,
         show_fig=False)
+
+def test_tabulate_fail(system_Haldane_PythTB):
+    k_nodes = [[0.0, 0.0, 0.5], [0.0, 0.0, 0.0], [0.5, 0.5, 0.5]]
+    path = wberri.Path(system_Haldane_PythTB, k_nodes=k_nodes, dk=1.0)
+
+    quantities = {
+        "Energy": wberri.calculators.tabulate.Energy(),
+        "berry": wberri.calculators.tabulate.BerryCurvature(),
+    }
+
+    key = "tabulate_grid"
+    calculators_fail = {"tabulate_grid": wberri.calculators.TabulatorAll(quantities,
+                                                                         ibands=[0],
+                                                                         mode="grid"),
+                        "ahc": wberri.calculators.static.AHC(Efermi=np.array([0,1,2]))}
+    for key, val in calculators_fail.items():
+        with pytest.raises(ValueError,
+                           match=f"Calculation along a Path is running, but calculator `{key}` is not compatible with a Path"):
+            wberri.run(system=system_Haldane_PythTB, grid=path, calculators={key: val})
